[JCR] Move history within its own namepace
* do not store gameModel histories in GM_<ID>/files/History any longer, but in GM_<ID>/history/
