+++ wegas-core/src/main/java/com/wegas/core/rest/FileController.java
@@ -8,9 +8,11 @@
 package com.wegas.core.rest;
 
 import com.wegas.core.ejb.GameModelFacade;
+import com.wegas.core.ejb.JCRFacade;
 import com.wegas.core.exception.client.WegasErrorMessage;
 import com.wegas.core.exception.client.WegasRuntimeException;
 import com.wegas.core.jcr.content.*;
+import com.wegas.core.jcr.content.ContentConnector.WorkspaceType;
 import com.wegas.core.jcr.content.FileDescriptor;
 import com.wegas.core.rest.util.annotations.CacheMaxAge;
 import org.apache.shiro.SecurityUtils;
@@ -21,8 +23,6 @@ import org.xml.sax.SAXException;
 
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
-import javax.jcr.ItemExistsException;
-import javax.jcr.LoginException;
 import javax.jcr.PathNotFoundException;
 import javax.jcr.RepositoryException;
 import javax.ws.rs.*;
@@ -30,16 +30,13 @@ import javax.ws.rs.core.*;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.transform.TransformerException;
 import java.io.*;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.zip.GZIPInputStream;
 import java.util.zip.GZIPOutputStream;
 import java.util.zip.ZipOutputStream;
+import javax.inject.Inject;
 
 /**
  * @author Cyril Junod (cyril.junod at gmail.com)
@@ -61,10 +58,12 @@ public class FileController {
     @EJB
     private GameModelFacade gmFacade;
 
-    /**
-     *
-     */
-    private static final String FILENAME_REGEXP = "^(?:[\\p{L}[0-9]-_ ]|\\.)+$";
+    @Inject
+    private JCRFacade jcrFacade;
+
+    private ContentConnector getContentConnector(long gameModelId) throws RepositoryException {
+        return ContentConnector.getFilesConnector(gameModelId);
+    }
 
     /**
      * @param gameModelId
@@ -75,7 +74,9 @@ public class FileController {
      * @param file
      * @param details
      * @param force       ovveride
+     *
      * @return HTTP 200 if everything ok, 4xx otherwise
+     *
      * @throws RepositoryException
      */
     @POST
@@ -83,13 +84,13 @@ public class FileController {
     @Produces(MediaType.APPLICATION_JSON)
     @Path("{force: (force/)?}upload{directory : .*?}")
     public Response upload(@PathParam("gameModelId") Long gameModelId,
-                           @FormDataParam("name") String name,
-                           @FormDataParam("note") String note,
-                           @FormDataParam("description") String description,
-                           @PathParam("directory") String path,
-                           @FormDataParam("file") InputStream file,
-                           @FormDataParam("file") FormDataBodyPart details,
-                           @PathParam("force") String force) throws RepositoryException {
+            @FormDataParam("name") String name,
+            @FormDataParam("note") String note,
+            @FormDataParam("description") String description,
+            @PathParam("directory") String path,
+            @FormDataParam("file") InputStream file,
+            @FormDataParam("file") FormDataBodyPart details,
+            @PathParam("force") String force) throws RepositoryException {
 
         SecurityUtils.getSubject().checkPermission("GameModel:Edit:gm" + gameModelId);
         logger.debug("File name: {}", details.getContentDisposition().getFileName());
@@ -101,9 +102,9 @@ public class FileController {
         try {
             if (details.getContentDisposition().getFileName() == null
                     || details.getContentDisposition().getFileName().equals("")) {//Assuming an empty filename means a directory
-                detachedFile = this.createDirectory(gameModelId, name, path, note, description);
+                detachedFile = jcrFacade.createDirectory(gameModelId, WorkspaceType.FILES, name, path, note, description);
             } else {
-                detachedFile = this.createFile(gameModelId, name, path, details.getMediaType().toString(),
+                detachedFile = jcrFacade.createFile(gameModelId, WorkspaceType.FILES, name, path, details.getMediaType().toString(),
                         note, description, file, override);
             }
         } catch (final WegasRuntimeException ex) {
@@ -133,21 +134,22 @@ public class FileController {
      * @param name
      * @param request
      * @param range       partial content range
+     *
      * @return the requested file with http 20x, 4xx if something went wrong
      */
     @GET
     @Path("read{absolutePath : .*?}")
     @CacheMaxAge(time = 48, unit = TimeUnit.HOURS)
     public Response read(@PathParam("gameModelId") Long gameModelId,
-                         @PathParam("absolutePath") String name,
-                         @Context Request request,
-                         @HeaderParam("Range") String range) {
+            @PathParam("absolutePath") String name,
+            @Context Request request,
+            @HeaderParam("Range") String range) {
 
         logger.debug("Asking file (/{})", name);
         AbstractContentDescriptor fileDescriptor;
         // ContentConnector connector = null;
         Response.ResponseBuilder response = Response.status(404);
-        try (final ContentConnector connector = new ContentConnector(gameModelId)) {
+        try (final ContentConnector connector = this.getContentConnector(gameModelId)) {
 
             fileDescriptor = DescriptorFactory.getDescriptor(name, connector);
             if (!SecurityUtils.getSubject().isPermitted("GameModel:View:gm" + gameModelId)) {
@@ -222,7 +224,7 @@ public class FileController {
     @Path("meta{absolutePath : .*?}")
     @Produces(MediaType.APPLICATION_JSON)
     public AbstractContentDescriptor getMeta(@PathParam("gameModelId") Long gameModelId, @PathParam("absolutePath") String name) {
-        try (final ContentConnector connector = new ContentConnector(gameModelId)) {
+        try (final ContentConnector connector = this.getContentConnector(gameModelId)) {
             AbstractContentDescriptor descriptor = DescriptorFactory.getDescriptor(name, connector);
             if (!SecurityUtils.getSubject().isPermitted("GameModel:View:gm" + gameModelId)) {
                 if (descriptor.isInheritedPrivate()) {
@@ -242,6 +244,7 @@ public class FileController {
     /**
      * @param gameModelId
      * @param directory
+     *
      * @return list of directory content
      */
     @GET
@@ -251,27 +254,14 @@ public class FileController {
 
         SecurityUtils.getSubject().checkPermission("GameModel:Edit:gm" + gameModelId);
 
-        logger.debug("Asking listing for directory (/{})", directory);
-        try (final ContentConnector connector = new ContentConnector(gameModelId)) {
-            AbstractContentDescriptor dir = DescriptorFactory.getDescriptor(directory, connector);
-            if (!dir.exist() || dir instanceof FileDescriptor) {
-                return null;
-            } else if (dir instanceof DirectoryDescriptor) {
-                List<AbstractContentDescriptor> ret = ((DirectoryDescriptor) dir).list();
-                Collections.sort(ret, new ContentComparator());
-                return ret;
-            }
-        } catch (LoginException ex) {
-            logger.error(null, ex);
-        } catch (RepositoryException ex) {
-            logger.error(null, ex);
-        }
-        return new ArrayList<>();
+        return jcrFacade.listDirectory(gameModelId, ContentConnector.WorkspaceType.FILES, directory);
     }
 
     /**
      * @param gameModelId
+     *
      * @return xml repository export
+     *
      * @throws RepositoryException
      * @throws IOException
      */
@@ -282,7 +272,7 @@ public class FileController {
 
         SecurityUtils.getSubject().checkPermission("GameModel:Edit:gm" + gameModelId);
 
-        final ContentConnector connector = new ContentConnector(gameModelId);
+        final ContentConnector connector = this.getContentConnector(gameModelId);
         StreamingOutput out = new StreamingOutput() {
             @Override
             public void write(OutputStream output) throws IOException, WebApplicationException {
@@ -299,7 +289,9 @@ public class FileController {
 
     /**
      * @param gameModelId
+     *
      * @return gzipped XML repository export
+     *
      * @throws RepositoryException
      */
     @GET
@@ -309,7 +301,7 @@ public class FileController {
 
         SecurityUtils.getSubject().checkPermission("GameModel:Edit:gm" + gameModelId);
 
-        final ContentConnector connector = new ContentConnector(gameModelId);
+        final ContentConnector connector = this.getContentConnector(gameModelId);
         StreamingOutput out = new StreamingOutput() {
             @Override
             public void write(OutputStream output) throws IOException, WebApplicationException {
@@ -332,7 +324,9 @@ public class FileController {
 
     /**
      * @param gameModelId
+     *
      * @return ZIP repository export
+     *
      * @throws RepositoryException
      */
     @GET
@@ -341,7 +335,7 @@ public class FileController {
 
         SecurityUtils.getSubject().checkPermission("GameModel:Edit:gm" + gameModelId);
 
-        final ContentConnector connector = new ContentConnector(gameModelId);
+        final ContentConnector connector = this.getContentConnector(gameModelId);
         StreamingOutput out = new StreamingOutput() {
             @Override
             public void write(OutputStream output) throws IOException, WebApplicationException {
@@ -360,7 +354,9 @@ public class FileController {
      * @param gameModelId
      * @param file
      * @param details
+     *
      * @return imported repository elements
+     *
      * @throws RepositoryException
      * @throws IOException
      * @throws SAXException
@@ -372,14 +368,14 @@ public class FileController {
     @Consumes(MediaType.MULTIPART_FORM_DATA)
     @Produces(MediaType.APPLICATION_JSON)
     public List<AbstractContentDescriptor> importXML(@PathParam("gameModelId") Long gameModelId,
-                                                     @FormDataParam("file") InputStream file,
-                                                     @FormDataParam("file") FormDataBodyPart details)
+            @FormDataParam("file") InputStream file,
+            @FormDataParam("file") FormDataBodyPart details)
             throws RepositoryException, IOException, SAXException,
             ParserConfigurationException, TransformerException {
 
         SecurityUtils.getSubject().checkPermission("GameModel:Edit:gm" + gameModelId);
 
-        try (final ContentConnector connector = new ContentConnector(gameModelId)) {
+        try (final ContentConnector connector = this.getContentConnector(gameModelId)) {
             switch (details.getMediaType().getSubtype()) {
                 case "x-gzip":
                 case "gzip":
@@ -405,7 +401,9 @@ public class FileController {
      * @param gameModelId
      * @param absolutePath
      * @param force
+     *
      * @return the destroyed element or HTTP not modified
+     *
      * @throws WegasErrorMessage when deleting a non empty directory without
      *                           force=true
      */
@@ -413,39 +411,19 @@ public class FileController {
     @Path("{force: (force/)?}delete{absolutePath : .*?}")
     @Produces(MediaType.APPLICATION_JSON)
     public Object delete(@PathParam("gameModelId") Long gameModelId,
-                         @PathParam("absolutePath") String absolutePath,
-                         @PathParam("force") String force) {
+            @PathParam("absolutePath") String absolutePath,
+            @PathParam("force") String force) {
 
         SecurityUtils.getSubject().checkPermission("GameModel:Edit:gm" + gameModelId);
 
-        final Boolean recursive = !force.equals("");
-        logger.debug("Asking delete for node ({}), force {}", absolutePath, recursive);
-        try (final ContentConnector connector = new ContentConnector(gameModelId)) {
-            AbstractContentDescriptor descriptor = DescriptorFactory.getDescriptor(absolutePath, connector);
-            if (descriptor.exist()) {
-                descriptor.sync();
-                if (descriptor instanceof DirectoryDescriptor && ((DirectoryDescriptor) descriptor).isRootDirectory()) {
-                    return Response.notModified("Unable to erase Root Directory").build();
-                }
-                try {
-                    descriptor.delete(recursive);
-                } catch (ItemExistsException e) {
-                    throw WegasErrorMessage.error(absolutePath + " is not empty, preventing removal");
-                }
-                return descriptor;
-            } else {
-                return Response.notModified("Path" + absolutePath + " does not exist").build();
-            }
-        } catch (RepositoryException ex) {
-            logger.error("Really what append here ??", ex);
-        }
-        return null;
+        return jcrFacade.delete(gameModelId, ContentConnector.WorkspaceType.FILES, absolutePath, force);
     }
 
     /**
      * @param tmpDescriptor
      * @param gameModelId
      * @param absolutePath
+     *
      * @return up to date descriptor
      */
     @PUT
@@ -453,14 +431,14 @@ public class FileController {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     public AbstractContentDescriptor update(AbstractContentDescriptor tmpDescriptor,
-                                            @PathParam("gameModelId") Long gameModelId,
-                                            @PathParam("absolutePath") String absolutePath) {
+            @PathParam("gameModelId") Long gameModelId,
+            @PathParam("absolutePath") String absolutePath) {
 
         AbstractContentDescriptor descriptor;
 
         SecurityUtils.getSubject().checkPermission("GameModel:Edit:gm" + gameModelId);
 
-        try (final ContentConnector connector = new ContentConnector(gameModelId)) {
+        try (final ContentConnector connector = this.getContentConnector(gameModelId)) {
 
             descriptor = DescriptorFactory.getDescriptor(absolutePath, connector);
             descriptor.setNote(tmpDescriptor.getNote());
@@ -486,134 +464,10 @@ public class FileController {
 
         SecurityUtils.getSubject().checkPermission("GameModel:Delete:gm" + gameModelId);
 
-        try (final ContentConnector fileManager = new ContentConnector(gameModelId)) {
+        try (final ContentConnector fileManager = this.getContentConnector(gameModelId)) {
             fileManager.deleteRoot();
         } catch (RepositoryException ex) {
             logger.error(null, ex);
         }
-    }
-
-    /**
-     * @param gameModelId
-     * @param name
-     * @param path
-     * @param mediaType
-     * @param note
-     * @param description
-     * @param file
-     * @param override
-     * @return new FileDescriptor
-     * @throws RepositoryException
-     */
-    public FileDescriptor createFile(Long gameModelId, String name, String path, String mediaType,
-                                     String note, String description, InputStream file, final Boolean override) throws RepositoryException {
-
-        logger.debug("File name: {}", name);
-
-        Pattern pattern = Pattern.compile(FILENAME_REGEXP);
-        Matcher matcher = pattern.matcher(name);
-        if (name.equals("") || !matcher.matches()) {
-            throw WegasErrorMessage.error(name + " is not a valid filename.  Letters, numbers, whitespace or \".-_\" only.");
-        }
-        try (final ContentConnector connector = new ContentConnector(gameModelId)) {
-
-            AbstractContentDescriptor dir = DescriptorFactory.getDescriptor(path, connector);
-            FileDescriptor detachedFile = new FileDescriptor(name, path, connector);
-
-            if (!detachedFile.exist() || override) {                                        //Node should not exist
-                detachedFile.setNote(note == null ? "" : note);
-                detachedFile.setDescription(description);
-                //TODO : check allowed mime-types
-                try {
-                    detachedFile.setBase64Data(file, mediaType);
-                    logger.info(name + "(" + mediaType + ") uploaded");
-                    return detachedFile;
-                } catch (IOException ex) {
-                    logger.error("Error reading uploaded file :", ex);
-                    throw WegasErrorMessage.error("Error reading uploaded file");
-                }
-            } else {
-                throw WegasErrorMessage.error(detachedFile.getPath() + name + " already exists");
-            }
-        } catch (RepositoryException ex) {
-            ex.printStackTrace();
-            throw ex;
-        }
-    }
-
-    /**
-     * @param gameModelId
-     * @param name
-     * @param path
-     * @param note
-     * @param description
-     * @return the new directory
-     * @throws RepositoryException
-     */
-    public DirectoryDescriptor createDirectory(Long gameModelId, String name, String path, String note, String description) throws RepositoryException {
-
-        //logger.debug("Directory name: {}", name);
-        Pattern pattern = Pattern.compile(FILENAME_REGEXP);
-        Matcher matcher = pattern.matcher(name);
-        if (name.equals("") || !matcher.matches()) {
-            throw WegasErrorMessage.error(name + " is not a valid filename.");
-        }
-        ContentConnector connector = new ContentConnector(gameModelId);
-        AbstractContentDescriptor dir = DescriptorFactory.getDescriptor(path, connector);
-        if (dir.exist()) {                                                      // Directory has to exist
-            DirectoryDescriptor detachedFile = new DirectoryDescriptor(name, path, connector);
-
-            if (!detachedFile.exist()) {                                        // Node should not exist
-                detachedFile.setNote(note == null ? "" : note);
-                detachedFile.setDescription(description);
-                detachedFile.sync();
-                logger.info("Directory {} created at {}", detachedFile.getName(), detachedFile.getPath());
-                return detachedFile;
-            } else {
-                throw WegasErrorMessage.error(detachedFile.getPath() + name + " already exists");
-            }
-        } else {
-            throw WegasErrorMessage.error(path + " directory does not exist already exists");
-        }
-    }
-
-    /**
-     * @param gameModelId
-     * @param path
-     * @return true if the directory exists
-     * @throws RepositoryException
-     */
-    public boolean directoryExists(Long gameModelId, String path) throws RepositoryException {
-        ContentConnector connector = new ContentConnector(gameModelId);
-        AbstractContentDescriptor dir = DescriptorFactory.getDescriptor(path, connector);
-        return dir.exist();
-    }
-
-    /**
-     * @param gameModelId
-     * @param path
-     * @return the file content
-     * @throws WegasErrorMessage when the requested file doesn't exists
-     */
-    public InputStream getFile(Long gameModelId, String path) {
-        logger.debug("Asking file (/{})", path);
-
-        InputStream ret = null;
-        AbstractContentDescriptor fileDescriptor = null;
-        ContentConnector connector = null;
-        try {
-            connector = new ContentConnector(gameModelId);
-            fileDescriptor = DescriptorFactory.getDescriptor(path, connector);
-        } catch (PathNotFoundException e) {
-            logger.debug("Asked path does not exist: {}", e.getMessage());
-            throw WegasErrorMessage.error("Directory " + path + " doest not exist");
-        } catch (RepositoryException e) {
-            logger.error("Need to check those errors", e);
-        }
-        if (fileDescriptor instanceof FileDescriptor) {
-            ret = new BufferedInputStream(((FileDescriptor) fileDescriptor).getBase64Data(), 512);
-            connector.save();
-        }
-        return ret;
     }
 }
